+++ app/Http/Controllers/UserController.php
@@ -1,0 +1,132 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
+class UserController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //tidak bisa diakses bila hak akses kurang dari 1::onlysee
+        if( Auth::user()->privileges->users  < 1 ) {
+            $params = [
+                'status'    => 'warning',
+                'message'   => 'Maaf anda tidak memiliki akses untuk melihat halaman ini'
+            ];
+            return redirect('home')->with($params);
+        }
+
+        // penguraian data
+        $params = [
+            'data'  => User::all(),
+            'type'  => 'data',
+            'title' => 'User / Employee'
+        ];
+
+        return view('layout.user.data', $params);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //tidak bisa diakses bila hak akses kurang dari 2::canedit
+        if( Auth::user()->privileges->users  < 2 ) {
+            $params = [
+                'status'    => 'warning',
+                'message'   => 'Maaf anda tidak memiliki akses untuk melihat halaman ini'
+            ];
+            return redirect('home')->with($params);
+        }
+
+        $data = [
+            'name'              => null,
+            'email'             => null,
+            'email_verified_at' => null,
+            'password'          => null,
+            'username'          => null,
+            'phone'             => null,
+            'privilege'         => null,
+            'address'           => null,
+            'image'             => 'default.jpg'
+        ];
+
+        $data = (object) $data;
+
+        // penguraian data
+        $params = [
+            'data'  => $data,
+            'type'  => 'create',
+            'title' => 'Create User / Employee'
+        ];
+
+        return view('layout.user.input', $params);
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        //
+    }
+}