+++ resources/views/sidebar/profile.blade.php
@@ -1,5 +1,9 @@
 <div class="d-flex fs-6 mb-3">
-    <img src="https://erwinzilla.com/v2/uploads/images/users/1642948549.jpg" class="avatar rounded-circle" alt="Avatar">
+    @if(Auth::user()->image)
+        <img src="{{ asset('uploads/images/users/'.$row->image) }}" class="avatar rounded-circle" alt="Avatar">
+    @else
+        <img src="{{ asset('uploads/images/users/default.jpg') }}" class="avatar rounded-circle" alt="Avatar">
+    @endif
     <div class="mx-2">
         <b class="text-black">{{ ucwords(Auth::user()->name) }}</b><br>
         <span class="badge bg-{{ Auth::user()->privileges->color }} bg-opacity-25 text-{{ Auth::user()->privileges->color }}">{{ ucwords(Auth::user()->privileges->name) }}</span>