+++ resources/views/sidebar/quick.blade.php
@@ -1,0 +1,1 @@
+<a href="#" class="btn btn-primary w-100 text-white">@svg('heroicon-s-clipboard-document-list', 'icon') New Ticket</a>